requirements.txt
equirements.txt
cachetools==4.2.4
google-auth==2.2.1
google-auth-oauthlib==0.4.6
gspread==4.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.7.2
simple-term-menu==1.6.3
blessed==1.20.0
tabulate==0.9.0